requirements.txt
Flask==0.10.1
Flask-Cors==1.9.0
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
docker-py==0.5.0
gevent==1.0.1
greenlet==0.4.4
itsdangerous==0.24
msgpack-python==0.4.2
numpy==1.9.0
psutil==2.1.3
pymongo==2.7.2
pyzmq==13.1.0
requests==2.4.3
six==1.8.0
websocket-client==0.20.0
wsgiref==0.1.2
zerorpc==0.4.4
PyYAML==3.11
